+++ src/main/java/uk/gov/hmcts/reform/unspec/model/docmosis/common/Respondent.java
@@ -1,8 +1,9 @@
-package uk.gov.hmcts.reform.unspec.model.docmosis.sealedclaim;
+package uk.gov.hmcts.reform.unspec.model.docmosis.common;
 
 import lombok.Builder;
 import lombok.Data;
 import uk.gov.hmcts.reform.unspec.model.Address;
+import uk.gov.hmcts.reform.unspec.model.docmosis.sealedclaim.Representative;
 
 @Data
 @Builder(toBuilder = true)
